backend/pyproject.toml
[project]
name = "helm-backend"
version = "0.1.0"
description = "Helm Marine Platform - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    "stripe>=11.0.0",
    "supabase>=2.0.0",
    "pyjwt>=2.8.0",
    "email-validator>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "boto3>=1.35.0",
    "resend>=2.0.0",
    "posthog>=3.0.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "llama-index>=0.12.0",
    "openai>=1.50.0",
    "pgvector>=0.3.0",
    "typesense>=0.21.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
